Create a limit using a conditional to prevent crash
In my own stick, I have four buttons that only have a single LED each -- when it was pressed rapidly after a different button that had multiple LEDs (4), the program would crash and that's no fun for anyone.

To fix that I added an index limit when trying to access `self.bg_colors` because the lists were mismatched. This might be a temporary fix if the root of the mismatch is identified, but it is a nice redundancy anyways!